Refactor: change directory structure.
