--- src/main/java/org/server/MainServer/Parser/InputParser.java
@@ -1,7 +1,0 @@
-package org.server.MainServer.Parser;
-
-public class InputParser {
-    public InputParser(){
-
-    }
-}
