+++ src/main/java/org/server/Utils/ServerConfig.java
@@ -1,4 +1,4 @@
-package org.server;
+package org.server.Utils;
 
 public class ServerConfig {
     public static int TCP_CONNECTION_DEFAULT_PORT = 8080;
